pulled non-db things out
