+++ src/db/orm/user.rs
@@ -1,7 +1,7 @@
 use super::Context;
 use crate::db::schema::*;
 use crate::db::schema::{projects, submissions};
-use crate::types::{Project, Submission};
+use crate::db::orm::{Project, Submission};
 use diesel::pg::expression::dsl::any;
 use diesel::BelongingToDsl;
 use diesel::ExpressionMethods;
