+++ src/main.rs
@@ -22,7 +22,8 @@ extern crate slug;
 extern crate diesel_derive_enum;
 
 use crate::db::Connection;
-use crate::types::Context;
+use crate::db::orm::Context;
+use crate::db::orm;
 use crate::utils::Result;
 use rocket::response::content;
 use rocket::*;
@@ -52,7 +53,7 @@ fn handle_graphql_get(
     request: juniper_rocket::GraphQLRequest,
     database: Connection,
 ) -> juniper_rocket::GraphQLResponse {
-    let schema = types::create_schema();
+    let schema = orm::create_schema();
     let context = Context { database };
     request.execute(&schema, &context)
 }
@@ -62,7 +63,7 @@ fn handle_graphql_post(
     request: juniper_rocket::GraphQLRequest,
     database: Connection,
 ) -> juniper_rocket::GraphQLResponse {
-    let schema = types::create_schema();
+    let schema = orm::create_schema();
     let context = Context { database };
     request.execute(&schema, &context)
 }
