+++ src/resolvers/permission.rs
@@ -1,6 +1,6 @@
 use crate::db::schema::{permissions, user_roles};
 use crate::db::sql_types::*;
-use crate::types::{Permission, User, UserRole};
+use crate::db::orm::{Permission, User, UserRole};
 use crate::utils::*;
 use diesel::dsl::any;
 use diesel::prelude::*;
