+++ src/routes/users.rs
@@ -1,6 +1,7 @@
 use crate::db::Connection;
 use crate::resolvers;
-use crate::types::{LoginRequest, Token, User, UserRequest};
+use crate::db::orm::{LoginRequest, User, UserRequest};
+use crate::types::Token;
 use crate::utils::Result;
 use rocket::http::Header;
 use rocket::{get, post, Responder};
